scripts(free-space): Add switch to disable docker cleaning

+++ scripts/free-space.sh
@@ -20,7 +20,6 @@ else
 	)
 
 	sudo apt purge -yq \
-		containerd.io \
 		snapd \
 		kubectl \
 		podman \
@@ -51,8 +50,12 @@ else
 	sudo rm -rf "/usr/local/share/boost"
 	sudo rm -rf "$AGENT_TOOLSDIRECTORY"
 
-	sudo docker image prune --all --force
-	sudo docker builder prune -a
+	# We shouldn't remove docker & it's images when running from `package_updates` workflow.
+	if [ "${CLEAN_DOCKER_IMAGES-true}" = "true" ]; then
+		sudo docker image prune --all --force
+		sudo docker builder prune -a
+		sudo apt purge -yq containerd.io
+	fi
 
 	sudo apt autoremove -yq
 	sudo apt clean
